refactor: extract error handling logic to ErrorHandler class
